requirements.txt
-e git+https://github.com/embedly/embedly-python.git@79b4dc457672f96f6376d7fed79efd379f32175b#egg=Embedly-0.2.0-py2.6-dev
django-jsonfield==0.8.11
django-cache-machine==0.8
